hate_classifier/requirements.txt
arabic_reshaper==3.0.0
fastapi==0.115.0
notificationapi_python_server_sdk==2.0.1
protobuf==6.31.1
pydantic==2.11.7
python-dotenv==1.0.1
python-bidi==0.6.6
uvicorn==0.34.3
google-genai==1.7.0